master:feat(ui): 增加基金详情页净值及历史涨跌幅字段
新增基金详情页的净值相关字段（净值日期、当日净值、估算净值等）和历史涨跌幅字段（近一年、近一季、近一月、近一周），并优化输入框精度控制。同时将腾讯云敏感配置从 application.properties 迁移至环境变量，提升安全性。添加 .env.example 和环境配置说明文档，便于开发者快速设置。

+++ src/main/java/com/buxuesong/account/infrastructure/general/entity/FundInfo.java
@@ -29,7 +29,7 @@ public class FundInfo {
     private BigDecimal holdingProfit;
 
     @JsonProperty("fundCount")
-    private BigDecimal fundCount;
+    private int fundCount;
 
     @JsonProperty("todayAmt")
     private BigDecimal todayAmt;
@@ -105,11 +105,11 @@ public class FundInfo {
         this.holdingProfit = holdingProfit;
     }
 
-    public BigDecimal getFundCount() {
+    public int getFundCount() {
         return fundCount;
     }
 
-    public void setFundCount(BigDecimal fundCount) {
+    public void setFundCount(int fundCount) {
         this.fundCount = fundCount;
     }
